requirements.txt
streamlit
python-docx
pandas
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
